Add intersects handler
