+++ src/test/resources/intersects_function.xml
@@ -1,0 +1,20 @@
+<?xml version="1.0"?>
+<h:html xmlns="http://www.w3.org/2002/xforms" xmlns:odk="http://www.opendatakit.org/xforms"
+        xmlns:jr="http://openrosa.org/javarosa" xmlns:h="http://www.w3.org/1999/xhtml"
+        xmlns:orx="http://openrosa.org/xforms">
+    <h:head>
+        <h:title>Form with intersects</h:title>
+        <model>
+            <instance>
+                <data id="intersects">
+                    <q1/>
+                    <calc/>
+                </data>
+            </instance>
+            <bind nodeset="/data/calc" calculate="intersects(/data/q1)" type="string"/>
+        </model>
+    </h:head>
+    <h:body>
+        <input ref="/data/q1"/>
+    </h:body>
+</h:html>
